refactor(docs): move doc processors into their own files
